Spring Boot 3로 버전 변경

+++ src/main/java/kr/co/triphos/chat/repository/ChatFileInfoRepository.java
@@ -1,14 +1,10 @@
 package kr.co.triphos.chat.repository;
 
 import kr.co.triphos.chat.entity.ChatFileInfo;
-import kr.co.triphos.common.entity.FileInfo;
-import org.apache.ibatis.annotations.Param;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
 
-import java.util.List;
 import java.util.Optional;
 
-public interface ChatFileInfoRepository extends JpaRepository<ChatFileInfo, Long> {
+public interface ChatFileInfoRepository extends JpaRepository<ChatFileInfo, Integer> {
 	Optional<ChatFileInfo> findByFileIdx(int fileIdx);
 }
